# LAB_1_PROCUREMENT_AGENT/open_api_tools/procurement_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Retail Procurement API",
    "description": "Retail procurement management API for submitting and viewing purchase orders, tracking price changes, and managing staff approvals. Compatible with Watsonx Orchestrate and Swagger UI.",
    "version": "1.0.0"
  },
  "paths": {
    "/orders": {
      "get": {
        "summary": "View purchase order history",
        "description": "Retrieve the complete history of purchase orders, including product_name, supplier, price, quantity, purchase_date, staff_in_charge, approver, latest_price_change information.",
        "operationId": "getOrderHistory",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHistoryResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new purchase order",
        "description": "Submit details to add a new purchase order, including product, supplier, quantity, and staff information. \n        Calculates price change compared to previous order for the same product. \n        User need to provide all fields in OrderRequest model including product_name, supplier, price, quantity, purchase_date (YYYY-MM-DD), staff_in_charge, and approver.\n        Always ask user for every fields when creating adding a new order.",
        "operationId": "addOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OrderHistoryItem": {
        "properties": {
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "supplier": {
            "type": "string",
            "title": "Supplier"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "purchase_date": {
            "type": "string",
            "title": "Purchase Date"
          },
          "staff_in_charge": {
            "type": "string",
            "title": "Staff In Charge"
          },
          "approver": {
            "type": "string",
            "title": "Approver"
          },
          "latest_price_change": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Price Change",
            "default": "-"
          }
        },
        "type": "object",
        "required": [
          "product_name",
          "supplier",
          "price",
          "quantity",
          "purchase_date",
          "staff_in_charge",
          "approver"
        ],
        "title": "OrderHistoryItem"
      },
      "OrderHistoryResponse": {
        "properties": {
          "orders": {
            "items": {
              "$ref": "#/components/schemas/OrderHistoryItem"
            },
            "type": "array",
            "title": "Orders"
          }
        },
        "type": "object",
        "required": [
          "orders"
        ],
        "title": "OrderHistoryResponse"
      },
      "OrderRequest": {
        "properties": {
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "supplier": {
            "type": "string",
            "title": "Supplier"
          },
          "price": {
            "type": "number",
            "title": "Price"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "purchase_date": {
            "type": "string",
            "title": "Purchase Date"
          },
          "staff_in_charge": {
            "type": "string",
            "title": "Staff In Charge"
          },
          "approver": {
            "type": "string",
            "title": "Approver"
          }
        },
        "type": "object",
        "required": [
          "product_name",
          "supplier",
          "price",
          "quantity",
          "purchase_date",
          "staff_in_charge",
          "approver"
        ],
        "title": "OrderRequest"
      },
      "OrderResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "latest_price_change": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Price Change",
            "default": "-"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "OrderResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://be-procurement-agent.1zy07nqib9k1.us-south.codeengine.appdomain.cloud",
      "description": "Production server"
    }
  ]
}